@@ -12,22 +12,35 @@ classifiers =
     Development Status :: 5 - Production/Stable
     Programming Language :: Python :: 3.6
 
-[aliases]
-test = pytest
-
 [options]
-packages = find:
+packages = flake8_alfred
 include_package_data = True
 
-python_requires = >=3.6
-install_requires = flake8
-setup_requires = pytest-runner
-tests_require = pytest
+python_requires = >=3.7
+install_requires = flake8>=3.5.0,<6.0.0
 
 [options.entry_points]
 flake8.extension = B1 = flake8_alfred:WarnSymbols
 
+[flake8]
+exclude = .venv
+
+[isort]
+force_single_line = True
+lexicographical = True
+lines_after_imports = 2
+src_paths =
+    flake8_alfred/
+    tests/
+
 [mypy]
+disallow_incomplete_defs = True
 disallow_untyped_defs = True
 ignore_missing_imports = True
-strict_optional = True
+warn_redundant_casts = True
+warn_return_any = True
+warn_unused_configs = True
+warn_unused_ignores = True
+
+[tool:pytest]
+addopts = "-vv"
